refactor(client): providers, DefaultLayout
