+++ packages/client/src/containers/DefaultLayout/DefaultLayout.styled.ts
