+++ packages/client/src/containers/DefaultLayout/DefaultLayout.types.d.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface DefaultLayoutProps {
+  children: ReactNode;
+}
