+++ packages/client/src/containers/index.ts
@@ -1,3 +1,4 @@
 export { default as Dashboard } from './Dashboard';
+export { default as DefaultLayout } from './DefaultLayout';
 export { default as SignIn } from './SignIn';
 export { default as SignUp } from './SignUp';
