+++ packages/client/src/elements/Inputs/TextField/index.tsx
@@ -20,7 +20,7 @@ const TextField = forwardRef<ITextFieldProps, TextFieldProps>(
     let internalEndAdornment = endAdornment;
 
     if (isPassword) {
-      internalStartAdornment = <LockIcon color={theme.palette.error} />;
+      internalStartAdornment = <LockIcon color={theme.palette.action.error} />;
       internalEndAdornment = <PasswordIconButton setPasswordMasked={setPasswordMasked} />;
     }
