+++ packages/client/src/hooks/useAuth.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { AuthContext } from '../providers/api/AuthProvider';
+import { AuthContext } from '../providers/api/AuthProvider/AuthProvider.context';
 
 const useAuth = () => useContext(AuthContext);
