+++ packages/client/src/hooks/useAxios.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { AxiosContext } from '../providers/api/AxiosProvider';
+import { AxiosContext } from '../providers/api/AxiosProvider/AxiosProvider.context';
 
 const useAxios = () => useContext(AxiosContext);
