+++ packages/client/src/hooks/useSnackbar.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { SnackbarContext } from '../providers/SnackbarProvider';
+import { SnackbarContext } from '../providers/SnackbarProvider/SnackbarProvider.context';
 import { ShowSnackbarProps } from '../providers/SnackbarProvider/SnackbarProvider.types';
 
 const useSnackbar = (): ((snackbarProps: ShowSnackbarProps) => void) => {
