+++ packages/client/src/hooks/useThemeType.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ThemeTypeContext } from '../providers/ThemeTypeProvider';
+import { ThemeTypeContext } from '../providers/ThemeTypeProvider/ThemeTypeProvider.context';
 import { ThemeTypeContextValue } from '../providers/ThemeTypeProvider/ThemeTypeProvider.types';
 
 const useThemeType = (): ThemeTypeContextValue => useContext(ThemeTypeContext);
