+++ packages/client/src/providers/SnackbarProvider/SnackbarProvider.context.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+import { SnackbarContextValue } from './SnackbarProvider.types';
+
+export const SnackbarContext = createContext<SnackbarContextValue>({
+  showSnackbar: () => {},
+});
