+++ packages/client/src/providers/StylesProvider/StylesProvider.types.d.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export interface StylesProviderProps {
+  children: ReactNode;
+}
