+++ packages/client/src/providers/ThemeTypeProvider/ThemeTypeProvider.context.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+
+import { ThemeType } from '../../theme/Theme';
+import { ThemeTypeContextValue } from './ThemeTypeProvider.types';
+
+export const ThemeTypeContext = createContext<ThemeTypeContextValue>({
+  themeType: ThemeType.dark,
+  setThemeType: () => null,
+  cleanThemeType: () => null,
+  setSystemThemeType: () => null,
+});
