+++ packages/client/src/providers/api/AuthProvider/AuthProvider.context.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { AuthProps } from './AuthProvider.types';
+
+export const AuthContext = createContext<AuthProps | undefined>(undefined);
