+++ packages/client/src/providers/api/AxiosProvider/AxiosProvider.context.ts
@@ -1,0 +1,4 @@
+import axios, { AxiosInstance } from 'axios';
+import { createContext } from 'react';
+
+export const AxiosContext = createContext<AxiosInstance>(axios);
