+++ packages/client/src/providers/index.tsx
@@ -1,11 +1,11 @@
 import React, { ReactNode } from 'react';
 import { I18nextProvider } from 'react-i18next';
 
+import { DefaultLayout } from '../containers';
 import i18n from '../i18n';
-import { StylesProvider } from '../theme';
-import AuthProvider from './api/AuthProvider';
-import AxiosProvider from './api/AxiosProvider';
+import { AuthProvider, AxiosProvider } from './api';
 import SnackbarProvider from './SnackbarProvider';
+import StylesProvider from './StylesProvider';
 
 interface ProvidersProps {
   children: ReactNode;
@@ -16,7 +16,9 @@ const Providers = ({ children }: ProvidersProps) => (
     <AxiosProvider>
       <AuthProvider>
         <I18nextProvider i18n={i18n}>
-          <SnackbarProvider>{children}</SnackbarProvider>
+          <SnackbarProvider>
+            <DefaultLayout>{children}</DefaultLayout>
+          </SnackbarProvider>
         </I18nextProvider>
       </AuthProvider>
     </AxiosProvider>
