--- packages/client/src/theme/index.ts
@@ -1,1 +1,0 @@
-export { default as StylesProvider } from '../providers/StylesProvider';
